Feat : 회원 수정 관련 API 구현 (미완성)

See merge request s09-fintech-finance-sub2/S09P22A405!101

+++ backend/src/main/java/com/ficrew/yourbutler/member/domain/entity/Member.java
@@ -22,9 +22,9 @@ public class Member {
     @Enumerated(EnumType.STRING)
     private Role role;
     private Integer age;
-    private Integer holdingAsset;
+    private Long holdingAsset;
     private Integer creditRating;
-    private Integer monthlyAvailableAsset;
+    private Long monthlyAvailableAsset;
     @Enumerated(EnumType.STRING)
     private NumberOfHouses numberOfHouses;
